+++ GoRogue.Snippets/HowTos/Factories/AdvancedFactory.cs
@@ -1,0 +1,37 @@
+﻿using GoRogue.Factories;
+using SadRogue.Primitives;
+// ReSharper disable NotAccessedPositionalProperty.Local
+// ReSharper disable UnusedVariable
+
+namespace GoRogue.Snippets.HowTos.Factories;
+
+#region AdvancedFactoryExample
+public static class AdvancedFactoryExample
+{
+    record Terrain(Point Position, int Glyph, bool IsWalkable, bool IsTransparent)
+        : IFactoryObject<string>
+    {
+        public string DefinitionID { get; set; } = "";
+    }
+
+    public static void ExampleCode()
+    {
+        // LambdaAdvancedFactoryBlueprint is the same as LambdaFactoryBlueprint but
+        // implements IAdvancedFactoryBlueprint instead, which allows its creation
+        // function to take parameters. This is useful, for example, to create objects
+        // that require parameters to be passed to their constructor.
+        var factory = new AdvancedFactory<string, Point, Terrain>
+        {
+            new LambdaAdvancedFactoryBlueprint<string, Point, Terrain>(
+                "Floor",
+                pos => new Terrain(pos, '.', true, true)),
+            new LambdaAdvancedFactoryBlueprint<string, Point, Terrain>(
+                "Wall",
+                pos => new Terrain(pos, '#', false, false))
+        };
+
+        var floorTile = factory.Create("Floor", new Point(1, 2));
+        var wallTile = factory.Create("Wall", new Point(3, 4));
+    }
+}
+#endregion